requirements.txt
faster-whisper>=1.2.1
PyAudioWPatch>=0.2.12.8
sounddevice>=0.5.1
numpy>=1.26
PyQt6>=6.7
Pillow>=10.4
SQLAlchemy>=2.0
requests>=2.32
pydantic>=2.7
PyYAML>=6.0
